feat(regression-testing): add performance tests
fix(sheepdog): use cdis-jenkins-lib master and fix google test (#62)

* fix(sheepdog): cdis-jenkins-lib master

* fix(sheepdog): remove console.log

* fix(googleServieAccountTests): add @

* fix(timeout): google retires

feat(query-tests): add performance tests for peregrine

feat(export performance tests): add export performance tests

fix(dcf-test): disable failing test (#65)

squash me

+++ codecept.conf.js
@@ -12,6 +12,8 @@ if (subdomain === 'default') {
 process.env.HOSTNAME = `${subdomain}.planx-pla.net`;
 console.log(`NAMESPACE: ${process.env.NAMESPACE}`);
 console.log(`HOSTNAME: ${process.env.HOSTNAME}`);
+console.log(`DATAURL: ${process.env.DATAURL}`);
+console.log(`TEST_DATA_PATH: ${process.env.TEST_DATA_PATH}`);
 
 exports.config = {
   output: './output',
@@ -64,11 +66,13 @@ exports.config = {
   bootstrap: './test_setup.js',
   teardown() {
     // session id is a global var retrieved in the helper
-    console.log(`Killing Selenium session ${seleniumSessionId}`);
-    request.del(`http://localhost:4444/wd/hub/session/${seleniumSessionId}`);
+    if ( typeof seleniumSessionId !== 'undefined' && seleniumSessionId ) {
+      console.log(`Killing Selenium session ${seleniumSessionId}`);
+      request.del(`http://localhost:4444/wd/hub/session/${seleniumSessionId}`);
+    }
   },
   hooks: [],
   tests: './suites/*/*.js',
-  timeout: 10000,
+  timeout: 30000,
   name: 'selenium',
 };
